Fix of sending GPS answer.

+++ conf/evolutions/default/2.sql
@@ -1,7 +1,7 @@
 # --- !Ups
 
 CREATE ALIAS IF NOT EXISTS geodistance AS $$
- double geodistance(double lat1, double lon1, double lat2, double lon2) {
+ int geodistance(double lat1, double lon1, double lat2, double lon2) {
       double Radius = 6372797.560856;;
       double dLat = Math.toRadians(lat2-lat1);;
       double dLon = Math.toRadians(lon2-lon1);;
@@ -9,7 +9,7 @@ CREATE ALIAS IF NOT EXISTS geodistance AS $$
          Math.cos(Math.toRadians(lat1)) * Math.cos(Math.toRadians(lat2)) *  
          Math.sin(dLon/2) * Math.sin(dLon/2);;
       double c = 2 * Math.asin(Math.sqrt(a));; 
-      return Radius * c;;
+      return (int)(Radius * c);;
    }
 $$;
